backend_api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.7
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#

arrow==1.2.3
    # via isoduration
attrs==22.2.0
    # via jsonschema
bravado-core==5.17.1
    # via pyramid-swagger
cached-property==1.5.2
    # via fqdn
certifi==2022.12.7
    # via requests
charset-normalizer==2.1.1
    # via requests
fqdn==1.5.1
    # via jsonschema
hupper==1.11
    # via pyramid
idna==3.4
    # via
    #   jsonschema
    #   requests
importlib-metadata==6.0.0
    # via
    #   jsonschema
    #   pastedeploy
    #   plaster
importlib-resources==5.10.2
    # via jsonschema
isoduration==20.11.0
    # via jsonschema
jsonpointer==2.3
    # via jsonschema
jsonref==1.0.1
    # via bravado-core
jsonschema[format]==4.17.3
    # via
    #   bravado-core
    #   pyramid-swagger
    #   swagger-spec-validator
msgpack==1.0.4
    # via bravado-core
mysql-connector-python==8.0.31
    # via -r requirements.in
pastedeploy==3.0.1
    # via plaster-pastedeploy
pkgutil-resolve-name==1.3.10
    # via jsonschema
plaster==1.1.2
    # via
    #   plaster-pastedeploy
    #   pyramid
plaster-pastedeploy==1.0.1
    # via pyramid
protobuf==3.20.1
    # via mysql-connector-python
pyramid==2.0
    # via
    #   -r requirements.in
    #   pyramid-swagger
pyramid-swagger==2.7.0
    # via -r requirements.in
pyrsistent==0.19.3
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   arrow
    #   bravado-core
pytz==2022.7
    # via bravado-core
pyyaml==6.0
    # via
    #   bravado-core
    #   swagger-spec-validator
requests==2.28.1
    # via bravado-core
rfc3339-validator==0.1.4
    # via jsonschema
rfc3987==1.3.8
    # via jsonschema
simplejson==3.18.1
    # via
    #   bravado-core
    #   pyramid-swagger
six==1.16.0
    # via
    #   bravado-core
    #   python-dateutil
    #   rfc3339-validator
swagger-spec-validator==3.0.3
    # via bravado-core
translationstring==1.4
    # via pyramid
typing-extensions==4.4.0
    # via
    #   arrow
    #   importlib-metadata
    #   jsonschema
    #   swagger-spec-validator
uri-template==1.2.0
    # via jsonschema
urllib3==1.26.13
    # via requests
uwsgi==2.0.18
    # via -r requirements.in
venusian==3.0.0
    # via pyramid
webcolors==1.12
    # via jsonschema
webob==1.8.7
    # via pyramid
zipp==3.11.0
    # via
    #   importlib-metadata
    #   importlib-resources
zope-deprecation==4.4.0
    # via pyramid
zope-interface==5.5.2
    # via pyramid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
